refactor(ingest): split IngestService into fact/mention/link path services (max-params 2/3)
IngestService had 8 injected deps and 1002 lines — the largest god-class in
the max-params=3 program. Split per the "write-path vs mention-path vs
link-path" brief into focused services, each ≤3 deps:

- EntityUpsertService (entityResolver?) — entity resolution shared by all paths
- FactEmbeddingService (embedder, hype) — embed/embedMany + HyPE alt-embedding
- FactResolverService (factEmbedding, predicateRegistry, metrics?) — the single
  fn::resolve_fact gateway: policy lookup, locale, embed, resolve, alt-embed
- FactIngestService (surreal, entityUpsert, factResolver) — typed ingestFact
- MentionExtractionService (extractor, factEmbedding) — extract + batch-embed
- MentionPersistService (surreal, entityUpsert, factResolver) — entity/fact/edge
- MentionIngestService (mentionExtraction, mentionPersist, metrics?) — orchestrate
- LinkIngestService (surreal, entityUpsert) — ingestLink / identity merge

IngestService is now a 42-line facade over factIngest/mentionIngest/linkIngest.
Public API (ingestFact/ingestMention/ingestLink) and behaviour unchanged.
Conflict weights read via process.env to keep FactResolverService at ≤3 deps.

tsc + typecheck + lint clean; 700 unit + 128 e2e + 9 jobs green.
